Mass conversion of i16 to f32
Integers up to ±2^24 are exactly represented by f32, so using i16 onyl adds hassles due to Vector2 conversions.

+++ soccer-fyrox/src/anchor.rs
@@ -5,5 +5,5 @@ pub enum Anchor {
     Center,
     // TopLeft is equivalent to a Custom anchored at (0,0).
     TopLeft,
-    Custom(Vector2<i16>),
+    Custom(Vector2<f32>),
 }
