+++ soccer-fyrox/src/main.rs
@@ -48,19 +48,19 @@ pub mod prelude {
     pub use crate::team::Team;
     pub use soccer_macros_fyrox::my_actor_based;
 
-    pub const WIDTH: i16 = 800;
-    pub const HEIGHT: i16 = 480;
+    pub const WIDTH: f32 = 800.;
+    pub const HEIGHT: f32 = 480.;
 
-    pub const HALF_WINDOW_W: i16 = WIDTH / 2;
+    pub const HALF_WINDOW_W: f32 = WIDTH / 2.;
 
     //# Size of level, including both the pitch and the boundary surrounding it
-    pub const LEVEL_W: i16 = 1000;
-    pub const LEVEL_H: i16 = 1400;
-    pub const HALF_LEVEL_W: i16 = LEVEL_W / 2;
-    pub const HALF_LEVEL_H: i16 = LEVEL_H / 2;
+    pub const LEVEL_W: f32 = 1000.;
+    pub const LEVEL_H: f32 = 1400.;
+    pub const HALF_LEVEL_W: f32 = LEVEL_W / 2.;
+    pub const HALF_LEVEL_H: f32 = LEVEL_H / 2.;
 
-    pub const HALF_PITCH_W: i16 = 442;
-    pub const HALF_PITCH_H: i16 = 622;
+    pub const HALF_PITCH_W: f32 = 442.;
+    pub const HALF_PITCH_H: f32 = 622.;
 
     // The below are specific to the port; drawing sequence doesn't work in 3d-based engines; actually,
     // using z-depth is more convenient, since draw calls can happen in any order.
