+++ soccer-fyrox/src/my_actor.rs
@@ -7,13 +7,13 @@ use crate::{anchor::Anchor, media::Media};
 //# Actor's x/y or pos attributes. When the object is drawn, we set self.pos (equivalent to setting both self.x and
 //# self.y) based on vpos, but taking scrolling into account.
 pub trait MyActor {
-    fn vpos(&self) -> Vector2<i16>;
+    fn vpos(&self) -> Vector2<f32>;
     fn img_base(&self) -> &'static str;
     fn img_indexes(&self) -> &[u8];
     fn anchor(&self) -> Anchor;
 
     //# We draw with the supplied offset to enable scrolling
-    fn draw(&self, scene: &mut Scene, media: &mut Media, offset_x: i16, offset_y: i16, z: f32) {
+    fn draw(&self, scene: &mut Scene, media: &mut Media, offset_x: f32, offset_y: f32, z: f32) {
         let pos_x = self.vpos().x - offset_x;
         let pos_y = self.vpos().y - offset_y;
